refactor: change page names
