+++ src/routes/flightRoutes.ts
@@ -1,6 +1,6 @@
 import { Router, Request, Response } from 'express';
 
-import { FlightController } from '../controller/FlightController';
+import { FlightController } from '../controllers/FlightController';
 
 const flightController = new FlightController();
